+++ apps/api/src/routes/v1_keys_updateKey.happy.test.ts
@@ -90,10 +90,9 @@ test("update all", async (t) => {
   expect(found?.name).toEqual("newName");
   expect(found?.ownerId).toEqual("newOwnerId");
   expect(found?.meta).toEqual(JSON.stringify({ new: "meta" }));
-  expect(found?.ratelimitType).toEqual("fast");
+  expect(found?.ratelimitAsync).toEqual(true);
   expect(found?.ratelimitLimit).toEqual(10);
-  expect(found?.ratelimitRefillRate).toEqual(5);
-  expect(found?.ratelimitRefillInterval).toEqual(1000);
+  expect(found?.ratelimitDuration).toEqual(1000);
   expect(found?.remaining).toEqual(0);
 });
 
@@ -139,10 +138,9 @@ test("update ratelimit", async (t) => {
   expect(found?.name).toEqual("test");
   expect(found?.ownerId).toBeNull();
   expect(found?.meta).toBeNull();
-  expect(found?.ratelimitType).toEqual("fast");
+  expect(found?.ratelimitAsync).toEqual(true);
   expect(found?.ratelimitLimit).toEqual(10);
-  expect(found?.ratelimitRefillRate).toEqual(5);
-  expect(found?.ratelimitRefillInterval).toEqual(1000);
+  expect(found?.ratelimitDuration).toEqual(1000);
   expect(found?.remaining).toBeNull();
 });
 
@@ -227,10 +225,9 @@ test("update should not affect undefined fields", async (t) => {
   expect(found?.ownerId).toEqual("newOwnerId");
   expect(found?.meta).toBeNull();
   expect(found?.expires).toEqual(key.expires);
-  expect(found?.ratelimitType).toBeNull();
+  expect(found?.ratelimitAsync).toBeNull();
   expect(found?.ratelimitLimit).toBeNull();
-  expect(found?.ratelimitRefillRate).toBeNull();
-  expect(found?.ratelimitRefillInterval).toBeNull();
+  expect(found?.ratelimitDuration).toBeNull();
   expect(found?.remaining).toBeNull();
 });