+++ apps/api/src/routes/v1_migrations_createKey.happy.test.ts
@@ -5,9 +5,11 @@ import { eq, schema } from "@unkey/db";
 import { newId } from "@unkey/id";
 import { IntegrationHarness } from "src/pkg/testutil/integration-harness";
 
+import type { ErrorResponse } from "@unkey/api/src";
 import { sha256 } from "@unkey/hash";
 import { KeyV1 } from "@unkey/keys";
 import type { V1KeysGetKeyResponse } from "./v1_keys_getKey";
+import type { V1KeysVerifyKeyRequest, V1KeysVerifyKeyResponse } from "./v1_keys_verifyKey";
 import type {
   V1MigrationsCreateKeysRequest,
   V1MigrationsCreateKeysResponse,
@@ -188,6 +190,56 @@ describe("with prefix", () => {
 describe("roles", () => {
   test("connects the specified roles", async (t) => {
     const h = await IntegrationHarness.init(t);
+    const permissions = ["p1", "p2"];
+    await h.db.primary.insert(schema.permissions).values(
+      permissions.map((name) => ({
+        id: newId("test"),
+        name,
+        workspaceId: h.resources.userWorkspace.id,
+      })),
+    );
+
+    const root = await h.createRootKey([`api.${h.resources.userApi.id}.create_key`]);
+
+    const res = await h.post<V1MigrationsCreateKeysRequest, V1MigrationsCreateKeysResponse>({
+      url: "/v1/migrations.createKeys",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${root.key}`,
+      },
+      body: [
+        {
+          start: "start_",
+          plaintext: "plaintext",
+          apiId: h.resources.userApi.id,
+          permissions,
+        },
+      ],
+    });
+
+    expect(res.status, `expected 200, received: ${JSON.stringify(res)}`).toBe(200);
+
+    const key = await h.db.readonly.query.keys.findFirst({
+      where: (table, { eq }) => eq(table.id, res.body.keyIds[0]),
+      with: {
+        permissions: {
+          with: {
+            permission: true,
+          },
+        },
+      },
+    });
+    expect(key).toBeDefined();
+    expect(key!.permissions.length).toBe(2);
+    for (const p of key!.permissions!) {
+      expect(permissions).include(p.permission.name);
+    }
+  });
+});
+
+describe("permissions", () => {
+  test("connects the specified permissions", async (t) => {
+    const h = await IntegrationHarness.init(t);
     const roles = ["r1", "r2"];
     await h.db.primary.insert(schema.roles).values(
       roles.map((name) => ({
@@ -274,12 +326,12 @@ test("creates a key with environment", async (t) => {
   expect(key!.environment).toBe(environment);
 });
 
-test("creates 50 keys", async (t) => {
+test("creates 100 keys", async (t) => {
   const h = await IntegrationHarness.init(t);
 
   const root = await h.createRootKey([`api.${h.resources.userApi.id}.create_key`]);
 
-  const req = new Array(50).fill(null).map(
+  const req = new Array(100).fill(null).map(
     (_, i) =>
       ({
         start: i.toString(),
@@ -335,7 +387,7 @@ test("an error rolls back and does not create any keys", async (t) => {
   // add a duplicate
   req.push(req[0]);
 
-  const res = await h.post<V1MigrationsCreateKeysRequest, V1MigrationsCreateKeysResponse>({
+  const res = await h.post<V1MigrationsCreateKeysRequest, ErrorResponse>({
     url: "/v1/migrations.createKeys",
     headers: {
       "Content-Type": "application/json",
@@ -344,7 +396,8 @@ test("an error rolls back and does not create any keys", async (t) => {
     body: req,
   });
 
-  expect(res.status).toEqual(500);
+  expect(res.status).toEqual(409);
+  expect(res.body.error.code).toEqual("NOT_UNIQUE");
 
   for (let i = 0; i < req.length; i++) {
     const key = await h.db.readonly.query.keys.findFirst({
@@ -404,4 +457,42 @@ test("retrieves a key in plain text", async (t) => {
 
   expect(getKeyRes.status).toBe(200);
   expect(getKeyRes.body.plaintext).toEqual(key);
+});
+
+test("migrate and verify a key", async (t) => {
+  const h = await IntegrationHarness.init(t);
+
+  const root = await h.createRootKey([`api.${h.resources.userApi.id}.create_key`]);
+
+  const key = new KeyV1({ byteLength: 16, prefix: "test" }).toString();
+
+  const res = await h.post<V1MigrationsCreateKeysRequest, V1MigrationsCreateKeysResponse>({
+    url: "/v1/migrations.createKeys",
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: `Bearer ${root.key}`,
+    },
+    body: [
+      {
+        plaintext: key,
+        apiId: h.resources.userApi.id,
+      },
+    ],
+  });
+  expect(res.status, `expected 200, received: ${JSON.stringify(res)}`).toBe(200);
+  expect(res.body.keyIds.length).toEqual(1);
+
+  const verifyRes = await h.post<V1KeysVerifyKeyRequest, V1KeysVerifyKeyResponse>({
+    url: "/v1/keys.verifyKey",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: {
+      apiId: h.resources.userApi.id,
+      key,
+    },
+  });
+
+  expect(verifyRes.status).toBe(200);
+  expect(verifyRes.body.valid).toEqual(true);
 });