+++ apps/dashboard/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @unkey/web
 
+## 0.1.26
+
+### Patch Changes
+
+- @unkey/ratelimit@0.1.12
+
+## 0.1.25
+
+### Patch Changes
+
+- @unkey/ratelimit@0.1.11
+
+## 0.1.24
+
+### Patch Changes
+
+- @unkey/ratelimit@0.1.10
+
+## 0.1.23
+
+### Patch Changes
+
+- @unkey/ratelimit@0.1.9
+
 ## 0.1.22
 
 ### Patch Changes