+++ packages/hono/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @unkey/hono
 
+## 1.1.14
+
+### Patch Changes
+
+- Updated dependencies [69c88fa]
+  - @unkey/api@0.20.7
+
+## 1.1.13
+
+### Patch Changes
+
+- Updated dependencies [56ad960]
+  - @unkey/api@0.20.6
+
+## 1.1.12
+
+### Patch Changes
+
+- Updated dependencies [e4961c7]
+  - @unkey/api@0.20.5
+
+## 1.1.11
+
+### Patch Changes
+
+- Updated dependencies [3ddb1f1]
+  - @unkey/api@0.20.4
+
 ## 1.1.10
 
 ### Patch Changes